backend/requirements.txt
django-ninja~=0.17.0
Django~=4.0.4
typing~=3.7.4.3
simplejwt~=2.0.1
pydantic~=1.9.0
requests~=2.28.0
# mysql
mysqlclient==2.1.0
# sqlserver
# mssql-django==1.1.2
# pyodbc==4.0.32
django-celery-beat==2.3.0
django-celery-results==2.3.1
celery==5.2.7
pypinyin~=0.46.0
user-agents==2.2.0
django_redis==5.2.0
openpyxl==3.0.10
django-simple-captcha~=0.5.17
oss2==2.16.0